also remove temporary files when an exception occurs

+++ src/nl/nikhef/jgridstart/gui/ActionInstall.java
@@ -7,6 +7,7 @@ import java.util.Random;
 
 import javax.swing.AbstractAction;
 import javax.swing.JFrame;
+import javax.swing.JOptionPane;
 import javax.swing.Timer;
 
 import nl.nikhef.jgridstart.CertificatePair;
@@ -35,8 +36,25 @@ public class ActionInstall extends CertificateAction {
     }
 
     public void actionPerformed(ActionEvent e) {
+	// explain what'll happen in password dialog or option pane
+	String message = 
+	    "You are about to install the selected certificate into your \n" +
+	    "web browser, so that you can access protected websites. \n" +
+	    "Please close your web browser completely before proceeding.";
+	Object [] options = { "Install", "Cancel" };
+	int ret = JOptionPane.showOptionDialog(parent,
+		message,
+		"Install certificate into browser",
+		JOptionPane.OK_CANCEL_OPTION,
+		JOptionPane.INFORMATION_MESSAGE,
+		null, // no icon
+		options, options[0]);
+	if (ret!=JOptionPane.OK_OPTION) return;
+	
 	CertificatePair cert = selection.getCertificatePair();
-	// TODO explain what'll happen in password dialog or option pane
+	logger.finer("Action: "+getValue(NAME));
+	
+	TempFileWriter pkcs = null;
 	try {
 	    /*
 	    final TempFileWriter pkcs = new TempFileWriter("browser", ".p12");
@@ -60,19 +78,20 @@ public class ActionInstall extends CertificateAction {
 	    //char[] pw = new char[15];
 	    char[] pw = new char[1]; // TODO Illegal key size exception if larger password :(
 	    for (int i=0; i<pw.length; i++) pw[i] = (char)(random.nextInt(128-32)+32);
-	    TempFileWriter pkcs = new TempFileWriter("browser", ".p12");
+	    pkcs = new TempFileWriter("browser", ".p12");
 	    PasswordCache.getInstance().set(pkcs.getPath(), pw);
 	    boolean oldAsk = PasswordCache.getInstance().setAlwaysAskForEncrypt(false);
 	    cert.exportTo(new File(pkcs.getPath()));
 	    PasswordCache.getInstance().setAlwaysAskForEncrypt(oldAsk);
 	    // now install and cleanup
 	    GridCertInstallerFirefox.install(pkcs.getPath(), pw);
-	    pkcs.delete();
 	} catch (PasswordCancelledException e1) {
 	    // do nothing
 	} catch (Exception e1) {
 	    logger.severe("Error installing certificate "+cert+": "+e1);
 	    ErrorMessage.error(parent, "Install failed", e1);
+	} finally {
+	    if (pkcs!=null) pkcs.delete();
 	}
     }
 }
